+++ src/components/UDPFilters/UDPFilters.js
@@ -69,10 +69,6 @@ class UDPFilters extends React.Component {
     intl: PropTypes.object
   });
 
-  static defaultProps = {
-    activeFilters: {}
-  };
-
   static getDerivedStateFromProps(props, state) {
     const newState = {};
     const arr = [];
@@ -122,7 +118,7 @@ class UDPFilters extends React.Component {
   };
 
   renderCheckboxFilter = (key, closedByDefault = false) => {
-    const { activeFilters } = this.props;
+    const { activeFilters = {} } = this.props;
     const groupFilters = activeFilters[key] || [];
 
     return (
@@ -153,7 +149,7 @@ class UDPFilters extends React.Component {
   };
 
   renderTagsFilter = () => {
-    const { activeFilters } = this.props;
+    const { activeFilters = {} } = this.props;
     const tagFilters = activeFilters.tags || [];
 
     return (
@@ -185,7 +181,7 @@ class UDPFilters extends React.Component {
   };
 
   renderErrorCodesFilter = () => {
-    const { activeFilters } = this.props;
+    const { activeFilters = {} } = this.props;
     const errorCodesFilters = activeFilters.errorCodes || [];
 
     return (
@@ -217,7 +213,7 @@ class UDPFilters extends React.Component {
   };
 
   renderReportTypesFiler = () => {
-    const { activeFilters } = this.props;
+    const { activeFilters = {} } = this.props;
     const reportTypesFilters = activeFilters.reportTypes || [];
 
     return (