+++ src/components/views/UDPForm.js
@@ -53,10 +53,6 @@ class UDPForm extends React.Component {
     submitting: PropTypes.bool,
   };
 
-  static defaultProps = {
-    initialValues: {},
-  };
-
   constructor(props) {
     super(props);
 
@@ -157,7 +153,7 @@ class UDPForm extends React.Component {
   }
 
   renderLastMenu() {
-    const { initialValues } = this.props;
+    const { initialValues = {} } = this.props;
     const { confirmDelete } = this.state;
     const isEditing = initialValues && initialValues.id;
 
@@ -250,17 +246,20 @@ class UDPForm extends React.Component {
     );
   };
 
-  renderFormPaneHeader = () => (
-    <PaneHeader
-      firstMenu={this.renderFirstMenu()}
-      lastMenu={this.renderLastMenu()}
-      paneTitle={this.props.initialValues.id ? this.props.initialValues.label : <FormattedMessage id="ui-erm-usage.udp.form.createUDP" />}
-    />
-  );
+  renderFormPaneHeader = () => {
+    const { initialValues = {} } = this.props;
+    return (
+      <PaneHeader
+        firstMenu={this.renderFirstMenu()}
+        lastMenu={this.renderLastMenu()}
+        paneTitle={initialValues.id ? initialValues.label : <FormattedMessage id="ui-erm-usage.udp.form.createUDP" />}
+      />
+    );
+  };
 
   render() {
     const {
-      initialValues,
+      initialValues = {},
       handleSubmit,
       data: { aggregators, harvesterImpls },
     } = this.props;